Global 'customFetch' configuration parameter
