+++ packages/handlers/json-schema/src/index.ts
@@ -1,7 +1,7 @@
 import { PredefinedProxyOptions, StoreProxy } from '@graphql-mesh/store';
-import { GetMeshSourceOptions, Logger, MeshHandler, MeshPubSub, YamlConfig } from '@graphql-mesh/types';
+import { GetMeshSourceOptions, ImportFn, Logger, MeshHandler, MeshPubSub, YamlConfig } from '@graphql-mesh/types';
 import { JSONSchemaLoaderBundle, createBundle, getGraphQLSchemaFromBundle } from '@omnigraph/json-schema';
-import { getCachedFetch, readFileOrUrl } from '@graphql-mesh/utils';
+import { readFileOrUrl } from '@graphql-mesh/utils';
 import { getInterpolatedHeadersFactory } from '@graphql-mesh/string-interpolation';
 import { process } from '@graphql-mesh/cross-helpers';
 
@@ -11,22 +11,25 @@ export default class JsonSchemaHandler implements MeshHandler {
   private bundleStoreProxy: StoreProxy<JSONSchemaLoaderBundle>;
   private baseDir: string;
   private logger: Logger;
-  private fetch: typeof fetch;
+  private fetchFn: typeof fetch;
+  private importFn: ImportFn;
   private pubsub: MeshPubSub;
 
   constructor({
     name,
     config,
     baseDir,
-    cache,
     store,
     pubsub,
     logger,
+    importFn,
+    fetchFn,
   }: GetMeshSourceOptions<YamlConfig.Handler['jsonSchema']>) {
     this.name = name;
     this.config = config;
     this.baseDir = baseDir;
-    this.fetch = getCachedFetch(cache);
+    this.fetchFn = fetchFn;
+    this.importFn = importFn;
     this.bundleStoreProxy = store.proxy('jsonSchemaBundle', PredefinedProxyOptions.JsonWithoutValidation);
     this.pubsub = pubsub;
     this.logger = logger;
@@ -38,12 +41,13 @@ export default class JsonSchemaHandler implements MeshHandler {
       const headersFactory = getInterpolatedHeadersFactory(config.bundleHeaders);
       const bundle = await readFileOrUrl<JSONSchemaLoaderBundle>(config.bundlePath, {
         cwd: this.baseDir,
-        fetch: this.fetch,
+        fetch: this.fetchFn,
         logger: this.logger,
         headers: headersFactory({
           env: process.env,
         }),
         fallbackFormat: 'json',
+        importFn: this.importFn,
       });
       return bundle;
     } else {
@@ -52,7 +56,7 @@ export default class JsonSchemaHandler implements MeshHandler {
           ...config,
           operations: config.operations as any,
           cwd: this.baseDir,
-          fetch: this.fetch,
+          fetch: this.fetchFn,
           logger: this.logger,
         });
       });
@@ -63,7 +67,7 @@ export default class JsonSchemaHandler implements MeshHandler {
     const bundle = await this.getDereferencedBundle();
     const schema = await getGraphQLSchemaFromBundle(bundle, {
       cwd: this.baseDir,
-      fetch: this.fetch,
+      fetch: this.fetchFn,
       pubsub: this.pubsub,
       logger: this.logger,
       baseUrl: this.config.baseUrl,
