+++ packages/handlers/odata/yaml-config.graphql
@@ -30,10 +30,6 @@ type ODataHandler @md {
   Use $expand for navigation props instead of seperate HTTP requests (Default: false)
   """
   expandNavProps: Boolean
-  """
-  Custom Fetch
-  """
-  customFetch: Any
 }
 
 enum ODataBatchMethod {
