+++ packages/handlers/thrift/package.json
@@ -41,7 +41,6 @@
     "@graphql-mesh/string-interpolation": "0.3.0",
     "@graphql-mesh/utils": "0.35.7",
     "@graphql-tools/utils": "8.6.13",
-    "fetchache": "0.1.2",
     "graphql-scalars": "1.17.0",
     "url-join": "4.0.1",
     "thrift": "0.16.0",
