+++ packages/mergers/bare/package.json
@@ -30,7 +30,6 @@
   "dependencies": {
     "@graphql-mesh/types": "0.75.0",
     "@graphql-mesh/utils": "0.35.7",
-    "fetchache": "0.1.2",
     "@graphql-tools/schema": "8.3.14",
     "@graphql-tools/wrap": "8.4.20",
     "@graphql-tools/utils": "8.6.13",
