+++ packages/runtime/src/get-mesh.ts
@@ -1,29 +1,15 @@
-import {
-  GraphQLSchema,
-  GraphQLResolveInfo,
-  OperationTypeNode,
-  GraphQLObjectType,
-  print,
-  SelectionSetNode,
-  Kind,
-  isLeafType,
-  getNamedType,
-  getOperationAST,
-  DocumentNode,
-} from 'graphql';
+import { GraphQLSchema, getOperationAST, DocumentNode } from 'graphql';
 import { ExecuteMeshFn, GetMeshOptions, SubscribeMeshFn } from './types';
 import {
   MeshPubSub,
   KeyValueCache,
   RawSourceOutput,
   GraphQLOperation,
-  SelectionSetParamOrFactory,
-  SelectionSetParam,
   Logger,
   MeshTransform,
 } from '@graphql-mesh/types';
 
-import { MESH_CONTEXT_SYMBOL, MESH_API_CONTEXT_SYMBOL } from './constants';
+import { MESH_CONTEXT_SYMBOL } from './constants';
 import {
   applySchemaTransforms,
   groupTransforms,
@@ -33,21 +19,13 @@ import {
   printWithCache,
 } from '@graphql-mesh/utils';
 
-import { delegateToSchema, IDelegateToSchemaOptions, StitchingInfo, SubschemaConfig } from '@graphql-tools/delegate';
-import { BatchDelegateOptions, batchDelegateToSchema } from '@graphql-tools/batch-delegate';
-import { WrapQuery } from '@graphql-tools/wrap';
-import {
-  AggregateError,
-  isAsyncIterable,
-  isDocumentNode,
-  mapAsyncIterator,
-  memoize1,
-  parseSelectionSet,
-} from '@graphql-tools/utils';
+import { SubschemaConfig } from '@graphql-tools/delegate';
+import { AggregateError, isAsyncIterable, mapAsyncIterator, memoize1 } from '@graphql-tools/utils';
 import { enableIf, envelop, PluginOrDisabledPlugin, useExtendContext, useSchema } from '@envelop/core';
 import { OneOfInputObjectsRule, useExtendedValidation } from '@envelop/extended-validation';
+import { getInContextSDK } from './in-context-sdk';
 
-export interface MeshInstance<TMeshContext = any> {
+export interface MeshInstance {
   execute: ExecuteMeshFn;
   subscribe: SubscribeMeshFn;
   schema: GraphQLSchema;
@@ -56,7 +34,6 @@ export interface MeshInstance<TMeshContext = any> {
   pubsub: MeshPubSub;
   cache: KeyValueCache;
   logger: Logger;
-  meshContext: TMeshContext;
   plugins: PluginOrDisabledPlugin[];
   getEnveloped: ReturnType<typeof envelop>;
   sdkRequesterFactory: (globalContext: any) => (document: DocumentNode, variables?: any, operationContext?: any) => any;
@@ -77,7 +54,7 @@ const memoizedGetEnvelopedFactory = memoize1(function getEnvelopedFactory(plugin
   });
 });
 
-export async function getMesh<TMeshContext = any>(options: GetMeshOptions): Promise<MeshInstance<TMeshContext>> {
+export async function getMesh(options: GetMeshOptions): Promise<MeshInstance> {
   const rawSources: RawSourceOutput[] = [];
   const {
     pubsub = new PubSub(),
@@ -86,8 +63,8 @@ export async function getMesh<TMeshContext = any>(options: GetMeshOptions): Prom
     additionalEnvelopPlugins = [],
     sources,
     merger,
-    additionalResolvers,
-    additionalTypeDefs,
+    additionalResolvers = [],
+    additionalTypeDefs = [],
     transforms,
   } = options;
 
@@ -151,190 +128,18 @@ export async function getMesh<TMeshContext = any>(options: GetMeshOptions): Prom
     transforms,
   });
 
-  getMeshLogger.debug(`Building Mesh Context`);
-  const meshContext: Record<string, any> = {
-    pubsub,
-    cache,
-    logger,
-    [MESH_CONTEXT_SYMBOL]: true,
-  };
-  getMeshLogger.debug(`Attaching in-context SDK, pubsub and cache to the context`);
-  const sourceMap = unifiedSchema.extensions.sourceMap as Map<RawSourceOutput, GraphQLSchema>;
-  await Promise.all(
-    rawSources.map(async rawSource => {
-      const rawSourceLogger = logger.child(`${rawSource.name}`);
-
-      const rawSourceContext: any = {
-        rawSource,
-        [MESH_API_CONTEXT_SYMBOL]: true,
-      };
-      // TODO: Somehow rawSource reference got lost in somewhere
-      let rawSourceSubSchemaConfig: SubschemaConfig;
-      const stitchingInfo = unifiedSchema.extensions.stitchingInfo as StitchingInfo;
-      if (stitchingInfo) {
-        for (const [subschemaConfig, subschema] of stitchingInfo.subschemaMap) {
-          if ((subschemaConfig as any).name === rawSource.name) {
-            rawSourceSubSchemaConfig = subschema;
-            break;
-          }
-        }
-      } else {
-        rawSourceSubSchemaConfig = rawSource;
-      }
-      const transformedSchema = sourceMap.get(rawSource);
-      const rootTypes: Record<OperationTypeNode, GraphQLObjectType> = {
-        query: transformedSchema.getQueryType(),
-        mutation: transformedSchema.getMutationType(),
-        subscription: transformedSchema.getSubscriptionType(),
-      };
-
-      rawSourceLogger.debug(`Generating In Context SDK`);
-      for (const operationType in rootTypes) {
-        const rootType: GraphQLObjectType = rootTypes[operationType];
-        if (rootType) {
-          rawSourceContext[rootType.name] = {};
-          const rootTypeFieldMap = rootType.getFields();
-          for (const fieldName in rootTypeFieldMap) {
-            const rootTypeField = rootTypeFieldMap[fieldName];
-            const inContextSdkLogger = rawSourceLogger.child(`InContextSDK.${rootType.name}.${fieldName}`);
-            const namedReturnType = getNamedType(rootTypeField.type);
-            const shouldHaveSelectionSet = !isLeafType(namedReturnType);
-            rawSourceContext[rootType.name][fieldName] = ({
-              root,
-              args,
-              context,
-              info = {
-                fieldName,
-                fieldNodes: [],
-                returnType: namedReturnType,
-                parentType: rootType,
-                path: {
-                  typename: rootType.name,
-                  key: fieldName,
-                  prev: undefined,
-                },
-                schema: transformedSchema,
-                fragments: {},
-                rootValue: root,
-                operation: {
-                  kind: Kind.OPERATION_DEFINITION,
-                  operation: operationType as OperationTypeNode,
-                  selectionSet: {
-                    kind: Kind.SELECTION_SET,
-                    selections: [],
-                  },
-                },
-                variableValues: {},
-                cacheControl: {
-                  setCacheHint: () => {},
-                  cacheHint: {},
-                },
-              },
-              selectionSet,
-              key,
-              argsFromKeys,
-              valuesFromResults,
-            }: {
-              root: any;
-              args: any;
-              context: any;
-              info: GraphQLResolveInfo;
-              selectionSet: SelectionSetParamOrFactory;
-              key?: string;
-              argsFromKeys?: (keys: string[]) => any;
-              valuesFromResults?: (result: any, keys?: string[]) => any;
-            }) => {
-              inContextSdkLogger.debug(`Called with`, {
-                args,
-                key,
-              });
-              const commonDelegateOptions: IDelegateToSchemaOptions = {
-                schema: rawSourceSubSchemaConfig,
-                rootValue: root,
-                operation: operationType as OperationTypeNode,
-                fieldName,
-                context,
-                transformedSchema,
-                info,
-              };
-              // If there isn't an extraction of a value
-              if (typeof selectionSet !== 'function') {
-                commonDelegateOptions.returnType = rootTypeField.type;
-              }
-              if (shouldHaveSelectionSet) {
-                let selectionCount = 0;
-                for (const fieldNode of info.fieldNodes) {
-                  if (fieldNode.selectionSet != null) {
-                    selectionCount += fieldNode.selectionSet.selections.length;
-                  }
-                }
-                if (selectionCount === 0) {
-                  if (!selectionSet) {
-                    throw new Error(
-                      `You have to provide 'selectionSet' for context.${rawSource.name}.${rootType.name}.${fieldName}`
-                    );
-                  }
-                  commonDelegateOptions.info = {
-                    ...info,
-                    fieldNodes: [
-                      {
-                        ...info.fieldNodes[0],
-                        selectionSet: {
-                          kind: Kind.SELECTION_SET,
-                          selections: [
-                            {
-                              kind: Kind.FIELD,
-                              name: {
-                                kind: Kind.NAME,
-                                value: '__typename',
-                              },
-                            },
-                          ],
-                        },
-                      },
-                      ...info.fieldNodes.slice(1),
-                    ],
-                  };
-                }
-              }
-              if (key && argsFromKeys) {
-                const batchDelegationOptions = {
-                  ...commonDelegateOptions,
-                  key,
-                  argsFromKeys,
-                  valuesFromResults,
-                } as unknown as BatchDelegateOptions;
-                if (selectionSet) {
-                  const selectionSetFactory = normalizeSelectionSetParamOrFactory(selectionSet);
-                  const path = [fieldName];
-                  const wrapQueryTransform = new WrapQuery(path, selectionSetFactory, identical);
-                  batchDelegationOptions.transforms = [wrapQueryTransform as any];
-                }
-                return batchDelegateToSchema(batchDelegationOptions);
-              } else {
-                const regularDelegateOptions: IDelegateToSchemaOptions = {
-                  ...commonDelegateOptions,
-                  args,
-                };
-                if (selectionSet) {
-                  const selectionSetFactory = normalizeSelectionSetParamOrFactory(selectionSet);
-                  const path = [fieldName];
-                  const wrapQueryTransform = new WrapQuery(path, selectionSetFactory, valuesFromResults || identical);
-                  regularDelegateOptions.transforms = [wrapQueryTransform as any];
-                }
-                return delegateToSchema(regularDelegateOptions);
-              }
-            };
-          }
-        }
-      }
-      meshContext[rawSource.name] = rawSourceContext;
-    })
-  );
-
   const plugins: PluginOrDisabledPlugin[] = [
     useSchema(unifiedSchema),
-    useExtendContext(() => meshContext),
+    useExtendContext(() => ({
+      pubsub,
+      cache,
+      logger,
+      [MESH_CONTEXT_SYMBOL]: true,
+    })),
+    enableIf(additionalTypeDefs.length > 0 || additionalResolvers.length > 0, () => {
+      const inContextSDK = getInContextSDK(unifiedSchema, rawSources, logger);
+      return useExtendContext(() => inContextSDK);
+    }),
     enableIf(!!unifiedSchema.getDirective('oneOf'), () =>
       useExtendedValidation({
         rules: [OneOfInputObjectsRule],
@@ -437,38 +242,10 @@ export async function getMesh<TMeshContext = any>(options: GetMeshOptions): Prom
       return pubsub.publish('destroy', undefined);
     },
     logger,
-    meshContext: meshContext as TMeshContext,
     plugins,
     get getEnveloped() {
       return memoizedGetEnvelopedFactory(plugins) as ReturnType<typeof envelop>;
     },
     sdkRequesterFactory,
   };
-}
-
-function normalizeSelectionSetParam(selectionSetParam: SelectionSetParam) {
-  if (typeof selectionSetParam === 'string') {
-    return parseSelectionSet(selectionSetParam);
-  }
-  if (isDocumentNode(selectionSetParam)) {
-    return parseSelectionSet(print(selectionSetParam));
-  }
-  return selectionSetParam;
-}
-
-function normalizeSelectionSetParamOrFactory(
-  selectionSetParamOrFactory: SelectionSetParamOrFactory
-): (subtree: SelectionSetNode) => SelectionSetNode {
-  return function getSelectionSet(subtree: SelectionSetNode) {
-    if (typeof selectionSetParamOrFactory === 'function') {
-      const selectionSetParam = selectionSetParamOrFactory(subtree);
-      return normalizeSelectionSetParam(selectionSetParam);
-    } else {
-      return normalizeSelectionSetParam(selectionSetParamOrFactory);
-    }
-  };
-}
-
-function identical<T>(val: T): T {
-  return val;
 }
