+++ packages/types/package.json
@@ -28,7 +28,6 @@
     "graphql": "*"
   },
   "dependencies": {
-    "fetchache": "0.1.2",
     "@graphql-mesh/store": "0.8.16",
     "@graphql-tools/delegate": "8.7.11",
     "@graphql-tools/utils": "8.6.13",
