+++ packages/utils/package.json
@@ -41,9 +41,7 @@
     "@graphql-typed-document-node/core": "3.1.1",
     "@graphql-mesh/cross-helpers": "0.1.6",
     "@graphql-mesh/types": "0.75.0",
-    "cross-undici-fetch": "^0.4.0",
     "chalk": "4.1.2",
-    "fetchache": "0.1.2",
     "graphql-subscriptions": "2.0.0",
     "lodash.get": "4.4.2",
     "lodash.set": "4.3.2",
