+++ patches/soap+0.45.0.patch
@@ -1,9 +1,13 @@
 diff --git a/node_modules/soap/index.js b/node_modules/soap/index.js
-index a1f0252..a5f8d8f 100644
+index a1f0252..f175219 100644
 --- a/node_modules/soap/index.js
 +++ b/node_modules/soap/index.js
-@@ -1,3 +1,4 @@
+@@ -1,3 +1,8 @@
  "use strict";
  
  module.exports = require('./lib/soap');
++module.exports.default = module.exports;
++module.exports.default = module.exports;
++module.exports.default = module.exports;
++module.exports.default = module.exports;
 +module.exports.default = module.exports;
