+++ website/docs/generated-markdown/SoapHandler.generated.md
@@ -17,19 +17,15 @@ You can also provide `.js` or `.ts` file path that exports schemaHeaders as an o
 * `operationHeaders` (type: `JSON`) - JSON object representing the Headers to add to the runtime of the API calls only for operation during runtime
 * `includePorts` (type: `Boolean`) - If true, the ports defined in the WSDL will be represented as GraphQL-Type objects in the schema.
 The fields of the object will be the operations of the port.
-
 Most soap-endpoints only define one port; so including it in the schema will just be inconvenient.
 But if there are multiple ports with operations of the same name, you should set this option to true.
 Otherwise only one of the identical-named operations will be callable.
-
 default: false
 * `includeServices` (type: `Boolean`) - If true, the services defined in the WSDL will be represented as GraphQL-Type objects in the schema.
 The fields of the object will be the ports of the service (or the operation, dependent on 'includePorts').
-
 Most soap-endpoints only define one service; so including it in the schema will just be inconvenient.
 But if there are multiple services with operations of the same name, you should set this option to true.
 Otherwise only one of the identical-named operations will be callable.
-
 default: false
 * `selectQueryOrMutationField` (type: `Array of Object`) - Allows to explicitly override the default operation (Query or Mutation) for any SOAP operation: 
   * `service` (type: `String`, required)
